requirements.txt
# Aplicação
streamlit
python-dotenv
Pillow

# Bases de dados
psycopg2-binary
pymongo
chromadb

# Dados e APIs
pandas
requests

# Crawlers e extração
selenium
webdriver-manager
lxml
trafilatura
tqdm

# Embeddings (ChromaDB)
sentence-transformers
